requirements.txt
blinker==1.9.0
cachelib==0.13.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
Flask==3.1.3
Flask-Session==0.8.0
gunicorn==26.0.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
msgspec==0.21.1
packaging==26.2
python-dotenv==1.2.2
redis==7.4.0
requests==2.33.1
urllib3==2.6.3
Werkzeug==3.1.8
